review fix: make sum in AccountMeResponse long

+++ front-ui/src/main/java/ru/yandex/practicum/mybankfront/client/dto/AccountMeResponse.java
@@ -2,5 +2,5 @@ package ru.yandex.practicum.mybankfront.client.dto;
 
 import java.time.LocalDate;
 
-public record AccountMeResponse(String name, LocalDate birthdate, int sum) {
+public record AccountMeResponse(String name, LocalDate birthdate, long sum) {
 }
